Avoid class initialization to allow post instrumentalization (#60)
When a class is instrumented by e.g. jacoco and has a field with an initializer that calls a method in an instrumented class the generator fails when loading and initialize the class with 'A required class was missing while executing org.assertj:assertj-assertions-generator-maven-plugin:2.0.0:generate-assertions: org/jacoco/agent/rt/internal_b0d6a23/Offline'. From what i have seen it is not necessary to initialize the class.