Adds support for Compile Tests (#37)
I ended up closing #37 because I was worried I generated
incorrect code, thus I added the suggestion in #37. Additionally,
I found bugs in generation and fixed them.

I refactored the generation tests to use a common junit "Rule"
called `GenerationPathHandler` that handles using a temporary
file per test case so there is no worry of collusion of generated
files. The Rule handles making all paths relative to the path
generated and deletes the folder IFF the tests pass, otherwise
the folder is kept and the path is printed to the error console.