Last commit for supporting assertions generation for generic types before reverting.
The main problem is that there is no simple way of getting the generic information through TypeToken.
From debugging TypeToken actually has the information we need but it is not always exposed in their API, sometimes their toString() is what we expect but sometimes it is not.
An example of that is ParameterizedType, if the TypeToken runtime type is a ParameterizedType the toString() is what we need unless it is an inner class.
See ClassUtilTest.create_generic_type_declarationwhwre you end up having thing like java.util.List<org.assertj.assertions.generator.util.ClassUtilTest.org.assertj.assertions.generator.util.ClassUtilTest$Foo<Integer>> instead of java.util.List<org.assertj.assertions.generator.util.ClassUtilTest.Foo<Integer>>.

Even if toString were correct, it is not a good idea to rely on them as they are not API, what if they in a newer version ?